Fix player being able to clip into wall when hit by enemy.
The normal player movement would check if underworld doors were present,
if so, it would do an alternative hit test that would check for walls
immediately above the player, as well.

The shove code, which is generic to actors, did not perform this added
check, but worked originally because of the margins check, which no
longer works here because they're all now expanded.